New tests for zero dependency checking are added
