+++ Sources/VGSCollectSDK/Core/Collector/VGSCollect+extension.swift
